+++ packages/components-js/tests/vrt/playwright/specs/custom/carousel.vrt.ts
@@ -3,6 +3,7 @@ import { setupScenario } from '../../helpers';
 import {
   viewportWidth3XL,
   viewportWidth4XL,
+  viewportWidthM,
   viewportWidthXXL,
 } from '@porsche-design-system/shared/testing/playwright.vrt.config';
 
@@ -17,5 +18,16 @@ test.describe(component, async () => {
       await setupScenario(page, `/${component}`, viewportWidth);
       await expect(page.locator('#app')).toHaveScreenshot(`${component}-${viewportWidth}.png`);
     });
+  });
+});
+
+// Test pagination on touch device
+test.describe(component, async () => {
+  test.skip(({ browserName }) => browserName !== 'chromium');
+  test.use({ hasTouch: true });
+
+  test(`should have no visual regression on touch device for viewport ${viewportWidthM}`, async ({ page }) => {
+    await setupScenario(page, `/${component}`, viewportWidthM);
+    await expect(page.locator('#app')).toHaveScreenshot(`${component}-${viewportWidthM}-touch.png`);
   });
 });